requirements.txt
httpx==0.23.3
fastapi==0.92.0
pytest==7.2.1
pytest_asyncio==0.20.3
python-dotenv==0.21.1
starlette==0.25.0
uvicorn==0.20.0
databases~=0.7.0
aioredis~=2.0.1
redis~=4.5.1
asyncpg~=0.27.0
SQLAlchemy~=1.4.46
pydantic~=1.10.5
alembic~=1.9.4
psycopg2~=2.9.5
passlib~=1.7.4
asyncio~=3.4.3
email-validator~=1.3.1